# kurs-6/modul-3/swagger/docs/docs.json
{
  "info": {
    "title": "Todo-list API",
    "description": "Api for Todo-list",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "todo": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "paths": {
    "/api/todo": {
      "get": {
        "description": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Todo"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "createdAt": {
                            "type": "string"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            },
            "description": "Fetch all todo's"
          }
        },
        "summary": "Fetch all todo's"
      },
      "post": {
        "description": null,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add a todo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to add a todo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": false
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Add a todo"
      }
    },
    "/api/todo/{id}": {
      "delete": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of todo to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a todo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Todo"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Failed to delete a todo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": false
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}